+++ packages/library-base/src/commonMain/kotlin/io/realm/MutableRealm.kt
@@ -79,9 +79,10 @@ public interface MutableRealm : TypedRealm {
     ): RealmQuery<T>
 
     /**
-     * Deletes the object from the underlying Realm. Only live objects can be deleted.
+     * Deletes object or objects of a query or result.
      *
-     * Frozen objects can be converted using [MutableRealm.findLatest]:
+     * *NOTE:* Only live objects can be deleted. Frozen objects can be converted using
+     * [MutableRealm.findLatest]:
      *
      * ```
      * val frozenObj = realm.query<Sample>.first().find()
@@ -90,9 +91,10 @@ public interface MutableRealm : TypedRealm {
      * }
      * ```
      *
+     * @param the [RealmObject], [RealmQuery] or [RealmResults] to delete.
      * @throws IllegalArgumentException if the object is invalid, frozen or not managed by Realm.
      */
-    public fun <T : RealmObject> delete(obj: T)
+    public fun delete(deleteable: Deleteable)
 }
 
 /**