+++ packages/library-base/src/commonMain/kotlin/io/realm/query/RealmQuery.kt
@@ -16,6 +16,7 @@
 
 package io.realm.query
 
+import io.realm.Deleteable
 import io.realm.MutableRealm
 import io.realm.Realm
 import io.realm.RealmConfiguration
@@ -202,7 +203,7 @@ public interface RealmQuery<T : RealmObject> : RealmElementQuery<T> {
 /**
  * Query returning [RealmResults].
  */
-public interface RealmElementQuery<T : RealmObject> {
+public interface RealmElementQuery<T : RealmObject> : Deleteable {
 
     /**
      * Finds all objects that fulfill the query conditions and returns them in a blocking fashion.