+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/RealmModelLoweringExtension.kt
@@ -19,6 +19,7 @@ package io.realm.compiler
 import io.realm.compiler.FqNames.MODEL_OBJECT_ANNOTATION
 import io.realm.compiler.FqNames.REALM_MODEL_COMPANION
 import io.realm.compiler.FqNames.REALM_OBJECT_INTERNAL_INTERFACE
+import io.realm.compiler.Names.REALM_OBJECT_INTERNAL_DELETE
 import io.realm.compiler.Names.REALM_OBJECT_INTERNAL_EMIT_FROZEN_UPDATE
 import io.realm.compiler.Names.REALM_OBJECT_INTERNAL_FREEZE
 import io.realm.compiler.Names.REALM_OBJECT_INTERNAL_IS_FROZEN
@@ -48,6 +49,7 @@ import org.jetbrains.kotlin.ir.util.primaryConstructor
 import org.jetbrains.kotlin.platform.konan.isNative
 
 private val realmObjectInternalOverrides = setOf(
+    REALM_OBJECT_INTERNAL_DELETE,
     REALM_OBJECT_INTERNAL_FREEZE,
     REALM_OBJECT_INTERNAL_THAW,
     REALM_OBJECT_INTERNAL_REGISTER_FOR_NOTIFICATION,